Stylelint
The GitHub action for this repository was failing because it was
attempting to run stylelint but there was no stylelint configuration. I
noticed that this github pages site seems to be using this Just the Docs
theme:

https://github.com/pmarsceill/just-the-docs

So I copied the .stylelintrc.json from there into here in the hopes that
the build will turn green. Maybe this dot file didn't get copied when
this repository was being set up for the first time. Or maybe it wasn't
committed? If that's the case then you may need to remove yours before
pulling these changes.